setting license to agpl, source files are public
Added a simple way of customizing the static files in www/ without
needing to change the actual source code. The idea is to make a copy of
www/ and then edit the copy.  This should make it easier for someone to
customize the look and feel of the site, but not have to keep a fork of
the orginal source up to date.

Not quite all the way implemented yet.